Bring back `__init__.py`
Remove unnecessary code in `list_submodules` function.
